rename package names
* `polyfuse-daemon` -> `polyfuse`
* `polyfuse` -> `polyfuse-types`
