--- crates/polyfuse-daemon/src/lib.rs
@@ -1,19 +1,0 @@
-#![warn(clippy::todo, clippy::unimplemented)]
-
-mod conn;
-mod daemon;
-mod parse;
-mod request;
-mod session;
-mod util;
-mod write;
-
-pub mod op;
-pub mod reply;
-
-pub use crate::{
-    daemon::{Builder, Daemon},
-    op::Operation,
-    request::Request,
-    session::{CapabilityFlags, ConnectionInfo},
-};
