+++ crates/polyfuse-types/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "polyfuse-types"
+version = "0.1.0-dev"
+authors = [ "Yusuke Sasaki <yusuke.sasaki.nuem@gmail.com>" ]
+license = "MIT OR Apache-2.0"
+edition = "2018"
+readme = "README.md"
+
+[dependencies]
+futures = "0.3"
+paste = "1"
+
+[dev-dependencies]
+version-sync = "0.8"
