+++ crates/polyfuse/src/lib.rs
@@ -1,26 +1,19 @@
-#![doc(html_root_url = "https://docs.rs/polyfuse/0.4.0-dev")]
+#![warn(clippy::todo, clippy::unimplemented)]
 
-//! A FUSE (Filesystem in userspace) library for Rust.
+mod conn;
+mod daemon;
+mod parse;
+mod request;
+mod session;
+mod util;
+mod write;
 
-#![deny(
-    missing_docs,
-    missing_debug_implementations,
-    clippy::cast_lossless,
-    clippy::cast_possible_truncation,
-    clippy::cast_possible_wrap,
-    clippy::cast_precision_loss,
-    clippy::cast_sign_loss,
-    clippy::checked_conversions,
-    clippy::invalid_upcast_comparisons
-)]
-#![forbid(clippy::unimplemented, clippy::todo)]
-
-pub mod error;
 pub mod op;
 pub mod reply;
-pub mod types;
 
-#[test]
-fn test_html_root_url() {
-    version_sync::assert_html_root_url_updated!("src/lib.rs");
-}
+pub use crate::{
+    daemon::{Builder, Daemon},
+    op::Operation,
+    request::Request,
+    session::{CapabilityFlags, ConnectionInfo},
+};
