+++ crates/polyfuse/src/parse.rs
@@ -1,5 +1,5 @@
-use polyfuse::types::FileLock;
 use polyfuse_kernel::{self as kernel, fuse_opcode};
+use polyfuse_types::types::FileLock;
 use std::{convert::TryFrom, ffi::OsStr, fmt, io, marker::PhantomData, mem, os::unix::prelude::*};
 
 pub struct Request<'a> {
