+++ crates/polyfuse/src/session.rs
