+++ crates/polyfuse/src/util.rs
