+++ examples/basic/Cargo.toml
@@ -6,7 +6,7 @@ edition = "2018"
 
 [dependencies]
 polyfuse = { path = "../../crates/polyfuse" }
-polyfuse-daemon = { path = "../../crates/polyfuse-daemon" }
+polyfuse-types = { path = "../../crates/polyfuse-types" }
 
 anyhow = "1"
 async-std = { version = ">=1.7.0", features = [ "attributes" ] }
